requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.5
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
apscheduler==3.11.0
pytz==2025.2
gunicorn==23.0.0
python-dotenv==1.1.1
